Consider any class name format valid
This technically allows edge cases where the package (or class, for classes with no package) can start with L and be incorrectly matched, but that feels like a case of play stupid games, win stupid prizes. Classes should be packaged, and package names should be lower case.
Requires a much less hacky change to JarMapping to optimize.